amazon/requirements.txt
alembic==1.0.11
altgraph==0.16.1
APScheduler==3.6.1
Babel==2.7.0
baidu-aip==2.2.17.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cssselect==1.0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Faker==2.0.1
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SocketIO==4.2.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
lxml==4.4.0
Mako==1.0.14
MarkupSafe==1.1.1
numpy==1.17.0
openpyxl==2.6.3
pandas==0.25.1
pefile==2019.4.18
Pillow==6.1.0
PyInstaller==3.5
pymongo==3.8.0
PyMySQL==0.9.3
pyquery==1.4.0
pytesseract==0.2.6
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.9.0
python-socketio==4.3.0
pytz==2019.1
pywin32-ctypes==0.2.0
redis==3.3.2
requests==2.22.0
selenium==3.141.0
six==1.12.0
SQLAlchemy==1.3.6
text-unidecode==1.2
tzlocal==2.0.0
urllib3==1.25.3
Werkzeug==0.15.5
WTForms==2.2.1
xlwt==1.3.0
